Orchestration_Service/requirements.txt
grpcio
PyYAML
redis
protobuf